project_clusters/data_science_bootcamp/team_01/requirements.txt
numpy
httpx
pandas
xgboost
scikit-learn
python-dotenv
requests
bs4
